Cleanup of free-programming-books.md
Updated the link for http://www.mmds.org/.
Removed An Introduction to Statistics with Python as it was not free. 
Removed The Architecture of Open Source Applications as it was a duplicate. 
Removed How to Write Parallel Programs as the domain is currently for sale. 
Removed CryptoParty handbook as it doesn't exist. 
Removed Network Science book as it 404s.